Parameters: condense into separate module
