+++ psyneulink/core/globals/__init__.py
@@ -4,6 +4,7 @@ from . import environment
 from . import keywords
 from . import kvo
 from . import log
+from . import parameters
 from . import preferences
 from . import registry
 from . import utilities
@@ -14,16 +15,18 @@ from .environment import *
 from .keywords import *
 from .kvo import *
 from .log import *
+from .parameters import *
 from .preferences import *
 from .registry import *
 from .utilities import *
 
-__all__ = list(defaults.__all__)
+__all__ = list(context.__all__)
+__all__.extend(defaults.__all__)
+__all__.extend(environment.__all__)
 __all__.extend(keywords.__all__)
 __all__.extend(kvo.__all__)
 __all__.extend(log.__all__)
+__all__.extend(parameters.__all__)
 __all__.extend(preferences.__all__)
 __all__.extend(registry.__all__)
-__all__.extend(environment.__all__)
 __all__.extend(utilities.__all__)
-__all__.extend(context.__all__)
