+++ psyneulink/library/components/mechanisms/processing/objective/predictionerrormechanism.py
@@ -155,11 +155,11 @@ from typing import Iterable
 import numpy as np
 import typecheck as tc
 
-from psyneulink.core.components.component import Param
 from psyneulink.core.components.functions.function import PredictionErrorDeltaFunction
 from psyneulink.core.components.mechanisms.mechanism import Mechanism_Base
 from psyneulink.core.components.states.outputstate import OutputState
-from psyneulink.core.globals.keywords import INITIALIZING, PREDICTION_ERROR_MECHANISM, OUTCOME, SAMPLE, TARGET
+from psyneulink.core.globals.keywords import INITIALIZING, OUTCOME, PREDICTION_ERROR_MECHANISM, SAMPLE, TARGET
+from psyneulink.core.globals.parameters import Param
 from psyneulink.core.globals.preferences.componentpreferenceset import is_pref_set, kpReportOutputPref
 from psyneulink.core.globals.preferences.preferenceset import PreferenceEntry, PreferenceLevel, kwPreferenceSetName
 from psyneulink.core.globals.utilities import is_numeric
